Update the Twig engine to the top of its develop branch
