+++ packages/engine-twig/lib/engine_twig.js
@@ -1,3 +1,5 @@
+'use strict';
+
 /*
  * twig pattern engine for patternlab-node - v0.15.1 - 2015
  *
@@ -18,10 +20,11 @@
  *
  */
 
-"use strict";
-
-var Twig = require('twig');
-var twig = Twig.twig;
+const fs = require('fs-extra');
+const path = require('path');
+const process = require('process');
+const Twig = require('node-twig');
+const twig = Twig.renderFile;
 
 var engine_twig = {
   engine: Twig,
@@ -38,11 +41,25 @@ var engine_twig = {
 
   // render it
   renderPattern: function renderPattern(pattern, data) {
-    var result = twig({
-      data: pattern.extendedTemplate
-    }).render(data);
-
-    return result;
+    return Promise.resolve(
+      twig(
+        pattern.relPath,
+        {
+          root: path.relative(
+            __dirname,
+            path.resolve(process.cwd(), 'source', '_patterns')
+          ),
+          context: data,
+        },
+        (error, template) => {
+          if (error) {
+            console.log(error);
+          }
+          console.log(template);
+          return template;
+        }
+      )
+    );
   },
 
   // find and return any {% include 'template-name' %} within pattern
@@ -50,7 +67,7 @@ var engine_twig = {
     var matches = pattern.template.match(this.findPartialsRE);
     return matches;
   },
-  findPartialsWithStyleModifiers: function () {
+  findPartialsWithStyleModifiers: function() {
     // TODO: make the call to this from oPattern objects conditional on their
     // being implemented here.
     return [];
@@ -58,25 +75,52 @@ var engine_twig = {
 
   // returns any patterns that match {{> value(foo:"bar") }} or {{>
   // value:mod(foo:"bar") }} within the pattern
-  findPartialsWithPatternParameters: function () {
+  findPartialsWithPatternParameters: function() {
     // TODO: make the call to this from oPattern objects conditional on their
     // being implemented here.
     return [];
   },
-  findListItems: function (pattern) {
+  findListItems: function(pattern) {
     var matches = pattern.template.match(this.findListItemsRE);
     return matches;
   },
 
   // given a pattern, and a partial string, tease out the "pattern key" and
   // return it.
-  findPartial: function (partialString) {
+  findPartial: function(partialString) {
     //var partialKey = partialString.replace(this.findPartialsRE, '$1');
     var partial = partialString.match(this.findPartialKeyRE)[0];
     partial = partial.replace(/"/g, '');
 
     return partial;
-  }
+  },
+
+  spawnFile: function(config, fileName) {
+    const paths = config.paths;
+    const metaFilePath = path.resolve(paths.source.meta, fileName);
+    try {
+      fs.statSync(metaFilePath);
+    } catch (err) {
+      //not a file, so spawn it from the included file
+      const metaFileContent = fs.readFileSync(
+        path.resolve(__dirname, '..', '_meta/', fileName),
+        'utf8'
+      );
+      fs.outputFileSync(metaFilePath, metaFileContent);
+    }
+  },
+
+  /**
+   * Checks to see if the _meta directory has engine-specific head and foot files,
+   * spawning them if not found.
+   *
+   * @param {object} config - the global config object from core, since we won't
+   * assume it's already present
+   */
+  spawnMeta: function(config) {
+    this.spawnFile(config, '_00-head.twig');
+    this.spawnFile(config, '_01-foot.twig');
+  },
 };
 
 module.exports = engine_twig;
