+++ packages/engine-twig/package.json
@@ -1,10 +1,11 @@
 {
-  "name": "@pattern-lab/engine-twig",
+  "name": "@pattern-lab/patternengine-node-twig",
   "description": "The Twig engine for Pattern Lab / Node",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "main": "lib/engine_twig.js",
   "dependencies": {
-    "twig": "^0.9.5"
+    "fs-extra": "^0.30.0",
+    "node-twig": "^1.1.0"
   },
   "devDependencies": {},
   "keywords": [
@@ -16,9 +17,11 @@
     "Javascript",
     "Twig"
   ],
-  "repository":
-    "https://github.com/pattern-lab/patternlab-node/tree/master/packages/engine-twig",
-  "bugs": "https://github.com/pattern-lab/patternlab-node/issues",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/pattern-lab/patternengine-node-twig.git"
+  },
+  "bugs": "https://github.com/pattern-lab/patternengine-node-twig/issues",
   "author": "Brian Muenzenmeyer & Geoffrey Pursell",
   "license": "MIT",
   "scripts": {
