--- packages/patternengine-node-twig/.github/ISSUE_TEMPLATE.md
@@ -1,9 +1,0 @@
-I am using Pattern Lab Node `vX.X.X` on `Windows|Mac|Linux`, using the `grunt|gulp` configuration.
-
-##### Expected Behavior
-
-
-##### Actual Behavior
-
-
-##### Steps to Reproduce
