--- packages/patternengine-node-twig/.github/PULL_REQUEST_TEMPLATE.md
@@ -1,5 +1,0 @@
-<!-- **Please read the contribution guidelines first, and target the `dev` branch!** -->
-
-Addresses #
-
-Summary of changes:
