--- packages/patternengine-node-twig/README.md
@@ -1,5 +1,0 @@
-## The Twig engine for Pattern Lab / Node
-
-To install the Twig engine in your edition, `npm install patternengine-node-twig` should do the trick.
-
-Level of Support is more or less full. Partial calls and lineage hunting are supported. Twig does not support the mustache-specific syntax extensions, style modifiers and pattern parameters, because their use cases are addressed by the core Twig feature set.
