--- packages/patternengine-node-twig/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@pattern-lab/patternengine-node-twig",
-  "description": "The Twig engine for Pattern Lab / Node",
-  "version": "0.2.0",
-  "main": "lib/engine_twig.js",
-  "dependencies": {
-    "fs-extra": "^0.30.0",
-    "node-twig": "^1.1.0"
-  },
-  "devDependencies": {},
-  "keywords": [
-    "Pattern Lab",
-    "Atomic Web Design",
-    "Node",
-    "Grunt",
-    "Gulp",
-    "Javascript",
-    "Twig"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/pattern-lab/patternengine-node-twig.git"
-  },
-  "bugs": "https://github.com/pattern-lab/patternengine-node-twig/issues",
-  "author": "Brian Muenzenmeyer & Geoffrey Pursell",
-  "license": "MIT",
-  "scripts": {
-    "test": "grunt travis --verbose"
-  },
-  "engines": {
-    "node": ">=4.0"
-  }
-}
